# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Socio AI Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/hallazgos/estructurar": {
      "post": {
        "tags": [
          "hallazgos"
        ],
        "summary": "Post Estructurar Hallazgo",
        "operationId": "post_estructurar_hallazgo_api_hallazgos_estructurar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HallazgoEstructurarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/hallazgos/tiempo-briefing": {
      "post": {
        "tags": [
          "hallazgos"
        ],
        "summary": "Post Tiempo Briefing",
        "operationId": "post_tiempo_briefing_api_hallazgos_tiempo_briefing_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BriefingTiempoLogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/normativa/refresh-monthly": {
      "post": {
        "tags": [
          "normativa"
        ],
        "summary": "Post Refresh Monthly",
        "operationId": "post_refresh_monthly_api_normativa_refresh_monthly_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiMeta": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "ApiMeta"
      },
      "ApiResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ],
            "title": "Status",
            "default": "ok"
          },
          "data": {
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse"
      },
...
      "BriefingTiempoLogRequest": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "area_codigo": {
            "type": "string",
            "title": "Area Codigo"
          },
          "area_nombre": {
            "type": "string",
            "title": "Area Nombre"
          },
          "tiempo_manual_min": {
            "type": "number",
            "title": "Tiempo Manual Min",
            "default": 0.0
          },
          "tiempo_ai_min": {
            "type": "number",
            "title": "Tiempo Ai Min",
            "default": 0.0
          },
          "notas": {
            "type": "string",
            "title": "Notas",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "cliente_id",
          "area_codigo",
          "area_nombre"
        ],
        "title": "BriefingTiempoLogRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HallazgoEstructurarRequest": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "title": "Cliente Id"
          },
          "area_codigo": {
            "type": "string",
            "title": "Area Codigo"
          },
          "area_nombre": {
            "type": "string",
            "title": "Area Nombre"
          },
          "marco": {
            "type": "string",
            "title": "Marco"
          },
          "riesgo": {
            "type": "string",
            "title": "Riesgo"
          },
          "afirmaciones_criticas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Afirmaciones Criticas"
          },
          "etapa": {
            "type": "string",
            "title": "Etapa",
            "default": "ejecucion"
          },
          "condicion_detectada": {
            "type": "string",
            "title": "Condicion Detectada"
          },
          "monto_estimado": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monto Estimado"
          },
          "causa_preliminar": {
            "type": "string",
            "title": "Causa Preliminar",
            "default": ""
          },
          "efecto_preliminar": {
            "type": "string",
            "title": "Efecto Preliminar",
            "default": ""
          },
          "guardar_en_hallazgos": {
            "type": "boolean",
            "title": "Guardar En Hallazgos",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "cliente_id",
          "area_codigo",
          "area_nombre",
          "marco",
          "riesgo",
          "condicion_detectada"
        ],
        "title": "HallazgoEstructurarRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}